Improve library (#3)
* Improve build tools

* Redesign the API

The classes were overkill for some situations.
The rewrite enables feature configuration and other improvements.
Endianness can be configured statically or at runtime.
String encoding can be disabled if not required.
Add EncodedEndianness to configure endianness.
BinaryReader now stores position by value not reference.
Add BinaryWriter variant which can be passed in a reference to an existing buffer.
Add option to disable size validation when reading.
Add _ prefix to internal variables.
Improve numeric reading and writing.
Use reinterpret_cast.

* Add read_numeric_array function

This does bounds checking once for the whole array instead of once per value.

* Add reserve function to writer

If multiple write calls are going to be made this can be called to ensure the buffer is pre-allocated.

* Add stub generator

* Add mypy config

* Update dev requirements

* Fix mypy errors

* Update tests

* Remove unused import

* Add default value for BinaryReader position

* Add endianness

* Add tests